requirements.txt
beautifulsoup4==4.11.1
pandas==1.4.1
Pillow==9.4.0
pytesseract==0.3.10
selenium==4.9.1
